requirements.txt
openvino==2026.1.0
optimum[openvino]>=1.17.0
gradio>=4.0
transformers>=4.35.0
azure-cognitiveservices-speech>=1.35.0
pyannote.audio>=3.1.0
pedalboard
noisereduce>=3.0
librosa
soundfile
python-dotenv
